Admin Added

+++ src/main/java/com/cg/jobportal/repository/AdminRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cg.jobportal.entity.Admin;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Integer> {
+public interface AdminRepository extends JpaRepository<Admin, Long> {
 
 }
